Added Direction Meter Fixed a few bugs
git-svn-id: http://firstforge.wpi.edu/svn/repos/zomb/trunk@85 0141d0df-edca-4edd-a601-3402401f7f08

+++ ZomBdll/WPF/Controls/ZomBGLControl.cs
@@ -156,7 +156,7 @@ namespace System451.Communication.Dashboard.WPF.Controls
         {
             if ((value is Thickness))
             {
-                return ((Thickness)value).Left * (double.Parse(parameter.ToString()));
+                return (double)(((Thickness)value).Left * (double.Parse(parameter.ToString())));
             }
             if ((value is double || value is int))
             {
